+++ CHANGELOG.md
@@ -10,15 +10,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Individual diagnostics have been created for each sort order issue.
-- The new diagnostics are now placed on the member that is out of order.
+
+### Changed
+
+- Moved enums into `RoslynMemberSorter.Enums` namespace and matching folder
+
+### Deprecated
+
+- Diagnostic `RMS0001` which was reported on the namespace or type that declared the member out of order.
 
 ### Fixed
 
 - Changelog date from "month-name day, year" to "year-month-day".
 
-### Deprecated
+### Removed
 
-- Diagnostic `RMS0001` which was reported on the namespace or type that declared the member out of order.
+- Tossed `RoslynMemberSorter.Accessibility` in favor of using `Microsoft.CodeAnalysis.Accessibility`.
 
 ## [0.1.0-alpha] 2022-03-25